// gen/common_pb.ts
export class Location extends Message<Location> {
  /**
   * The file path
   *
   * @generated from field: optional string file = 1;
   */
  file?: string;

  /**
   * The Rust module path
   *
   * @generated from field: optional string module_path = 2;
   */
  modulePath?: string;

  /**
   * The line number in the source code file.
   *
   * @generated from field: optional uint32 line = 3;
   */
  line?: number;

  /**
   * The character in `line`.
   *
   * @generated from field: optional uint32 column = 4;
   */
  column?: number;

  constructor(data?: PartialMessage<Location>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.common.Location";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "file", kind: "scalar", T: 9 /* ScalarType.STRING */, opt: true },
    { no: 2, name: "module_path", kind: "scalar", T: 9 /* ScalarType.STRING */, opt: true },
    { no: 3, name: "line", kind: "scalar", T: 13 /* ScalarType.UINT32 */, opt: true },
    { no: 4, name: "column", kind: "scalar", T: 13 /* ScalarType.UINT32 */, opt: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Location {
    return new Location().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Location {
    return new Location().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Location {
    return new Location().fromJsonString(jsonString, options);
  }

  static equals(a: Location | PlainMessage<Location> | undefined, b: Location | PlainMessage<Location> | undefined): boolean {
    return proto3.util.equals(Location, a, b);
  }
}
// ...
export class Metadata extends Message<Metadata> {
  /**
   * The name of the span or event.
   *
   * @generated from field: string name = 1;
   */
  name = "";

  /**
   * Describes the part of the system where the span or event that this
   * metadata describes occurred.
   *
   * @generated from field: string target = 2;
   */
  target = "";

  /**
   * The path to the Rust module where the span occurred.
   *
   * @generated from field: string module_path = 3;
   */
  modulePath = "";

  /**
   * The Rust source location associated with the span or event.
   *
   * @generated from field: rs.tokio.console.common.Location location = 4;
   */
  location?: Location;

  /**
   * Indicates whether metadata is associated with a span or with an event.
   *
   * @generated from field: rs.tokio.console.common.Metadata.Kind kind = 5;
   */
  kind = Metadata_Kind.SPAN;

  /**
   * Describes the level of verbosity of a span or event.
   *
   * @generated from field: rs.tokio.console.common.Metadata.Level level = 6;
   */
  level = Metadata_Level.ERROR;

  /**
   * The names of the key-value fields attached to the
   * span or event this metadata is associated with.
   *
   * @generated from field: repeated string field_names = 7;
   */
  fieldNames: string[] = [];

  constructor(data?: PartialMessage<Metadata>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.common.Metadata";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "target", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "module_path", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "location", kind: "message", T: Location },
    { no: 5, name: "kind", kind: "enum", T: proto3.getEnumType(Metadata_Kind) },
    { no: 6, name: "level", kind: "enum", T: proto3.getEnumType(Metadata_Level) },
    { no: 7, name: "field_names", kind: "scalar", T: 9 /* ScalarType.STRING */, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Metadata {
    return new Metadata().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Metadata {
    return new Metadata().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Metadata {
    return new Metadata().fromJsonString(jsonString, options);
  }

  static equals(a: Metadata | PlainMessage<Metadata> | undefined, b: Metadata | PlainMessage<Metadata> | undefined): boolean {
    return proto3.util.equals(Metadata, a, b);
  }
}

/**
 * Indicates whether metadata is associated with a span or with an event.
 *
 * @generated from enum rs.tokio.console.common.Metadata.Kind
 */
export enum Metadata_Kind {
  /**
   * Indicates metadata is associated with a span.
   *
   * @generated from enum value: SPAN = 0;
   */
  SPAN = 0,

  /**
   * Indicates metadata is associated with an event.
   *
   * @generated from enum value: EVENT = 1;
   */
  EVENT = 1,
}
// Retrieve enum metadata with: proto3.getEnumType(Metadata_Kind)
proto3.util.setEnumType(Metadata_Kind, "rs.tokio.console.common.Metadata.Kind", [
  { no: 0, name: "SPAN" },
  { no: 1, name: "EVENT" },
]);

/**
 * Describes the level of verbosity of a span or event.
 *
 * Corresponds to `Level` in the `tracing` crate.
 *
 * @generated from enum rs.tokio.console.common.Metadata.Level
 */
export enum Metadata_Level {
  /**
   * The "error" level.
   *
   * Designates very serious errors.
   *
   * @generated from enum value: ERROR = 0;
   */
  ERROR = 0,

  /**
   * The "warn" level.
   *
   * Designates hazardous situations.
   *
   * @generated from enum value: WARN = 1;
   */
  WARN = 1,

  /**
   * The "info" level.
   * Designates useful information.
   *
   * @generated from enum value: INFO = 2;
   */
  INFO = 2,

  /**
   * The "debug" level.
   *
   * Designates lower priority information.
   *
   * @generated from enum value: DEBUG = 3;
   */
  DEBUG = 3,

  /**
   * The "trace" level.
   *
   * Designates very low priority, often extremely verbose, information.
   *
   * @generated from enum value: TRACE = 4;
   */
  TRACE = 4,
}