// gen/instrument_pb.ts
export class InstrumentRequest extends Message<InstrumentRequest> {
  constructor(data?: PartialMessage<InstrumentRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.instrument.InstrumentRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): InstrumentRequest {
    return new InstrumentRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): InstrumentRequest {
    return new InstrumentRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): InstrumentRequest {
    return new InstrumentRequest().fromJsonString(jsonString, options);
  }

  static equals(a: InstrumentRequest | PlainMessage<InstrumentRequest> | undefined, b: InstrumentRequest | PlainMessage<InstrumentRequest> | undefined): boolean {
    return proto3.util.equals(InstrumentRequest, a, b);
  }
}
// ...
export class Update extends Message<Update> {
  /**
   * The system time when this update was recorded.
   *
   * This is the timestamp any durations in the included `Stats` were
   * calculated relative to.
   *
   * @generated from field: google.protobuf.Timestamp now = 1;
   */
  now?: Timestamp;

  /**
   * Task state update.
   *
   * @generated from field: rs.tokio.console.tasks.TaskUpdate task_update = 2;
   */
  taskUpdate?: TaskUpdate;

  /**
   * Resource state update.
   *
   * @generated from field: rs.tokio.console.resources.ResourceUpdate resource_update = 3;
   */
  resourceUpdate?: ResourceUpdate;

  /**
   * Async operations state update
   *
   * @generated from field: rs.tokio.console.async_ops.AsyncOpUpdate async_op_update = 4;
   */
  asyncOpUpdate?: AsyncOpUpdate;

  /**
   * Any new span metadata that was registered since the last update.
   *
   * @generated from field: rs.tokio.console.common.RegisterMetadata new_metadata = 5;
   */
  newMetadata?: RegisterMetadata;

  constructor(data?: PartialMessage<Update>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.instrument.Update";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "now", kind: "message", T: Timestamp },
    { no: 2, name: "task_update", kind: "message", T: TaskUpdate },
    { no: 3, name: "resource_update", kind: "message", T: ResourceUpdate },
    { no: 4, name: "async_op_update", kind: "message", T: AsyncOpUpdate },
    { no: 5, name: "new_metadata", kind: "message", T: RegisterMetadata },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Update {
    return new Update().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Update {
    return new Update().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Update {
    return new Update().fromJsonString(jsonString, options);
  }

  static equals(a: Update | PlainMessage<Update> | undefined, b: Update | PlainMessage<Update> | undefined): boolean {
    return proto3.util.equals(Update, a, b);
  }
}